% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_bill.R
\name{cg_bill}
\alias{cg_bill}
\title{Get a specific bill}
\usage{
cg_bill(congress, id, key = NULL, as = "table", ...)
}
\arguments{
\item{congress}{(character) congress number, e.g. 110}

\item{id}{(character) a bill slug, for example hr4881 - these can be
found in the recent bill response.}

\item{key}{your SunlightLabs API key; loads from environment variable from .Renviron or
from an option from .Rprofile}

\item{as}{(character) One of table (default), list, or response (httr response object).
When table is requested, the default, a table is not always returned, but is returned almost
always. You can then make a table yourself if you like.}

\item{...}{Optional additional curl options (debugging tools mostly)
passed on to \code{\link[crul]{HttpClient}}}
}
\value{
some nasty parsing, so just list of lists for now
}
\description{
Get a specific bill
}
\examples{
\dontrun{
cg_bill(115, "hr21")
}
}
