% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_bill_subjects_search.R
\name{cg_bill_subjects_search}
\alias{cg_bill_subjects_search}
\title{Search for bill subjects that contain a specified term}
\usage{
cg_bill_subjects_search(query, key = NULL, as = "table", ...)
}
\arguments{
\item{query}{(character) a word or phrase to search}
}
\description{
Search for bill subjects that contain a specified term
}
\examples{
\dontrun{
cg_bill_subjects_search('climate')
}
}
