% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_bills_recent.R
\name{cg_bills_recent}
\alias{cg_bills_recent}
\title{Get recent bills}
\usage{
cg_bills_recent(congress, chamber, type, key = NULL, as = "table", ...)
}
\arguments{
\item{congress}{The number of the Congress this update took place during.}

\item{chamber}{The chamber this update took place in. 'house' or 'senate'}

\item{type}{Type of bill. introduced, updated, active, passed, enacted
or vetoed}

\item{key}{your SunlightLabs API key; loads from environment variable from .Renviron or
from an option from .Rprofile}

\item{as}{(character) One of table (default), list, or response (httr response object).
When table is requested, the default, a table is not always returned, but is returned almost
always. You can then make a table yourself if you like.}

\item{...}{Optional additional curl options (debugging tools mostly)
passed on to \code{\link[crul]{HttpClient}}}
}
\value{
some nasty parsing, so just list of lists for now
}
\description{
Get recent bills
}
\examples{
\dontrun{
cg_bills_recent(115, "house", "introduced")
}
}
