% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_bills_recent_member.R
\name{cg_bills_recent_member}
\alias{cg_bills_recent_member}
\title{Get recent bills by member id}
\usage{
cg_bills_recent_member(id, type, key = NULL, as = "table", ...)
}
\arguments{
\item{id}{(character) The ID of the member to retrieve; it is
assigned by the Biographical Directory of the United States
Congress (\url{http://bioguide.congress.gov/biosearch/biosearch.asp})
or can be retrieved from a \code{\link[=cg_members]{cg_members()}} request}

\item{type}{Type of bill. introduced or updated}

\item{key}{your SunlightLabs API key; loads from environment variable from .Renviron or
from an option from .Rprofile}

\item{as}{(character) One of table (default), list, or response (httr response object).
When table is requested, the default, a table is not always returned, but is returned almost
always. You can then make a table yourself if you like.}

\item{...}{Optional additional curl options (debugging tools mostly)
passed on to \code{\link[crul]{HttpClient}}}
}
\value{
some nasty parsing, so just list of lists for now
}
\description{
Get recent bills by member id
}
\examples{
\dontrun{
cg_bills_recent_member("L000287", "introduced")
}
}
