% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_bills_recent_subject.R
\name{cg_bills_recent_subject}
\alias{cg_bills_recent_subject}
\title{Get recent bills by subject}
\usage{
cg_bills_recent_subject(subject, key = NULL, as = "table", ...)
}
\arguments{
\item{subject}{(character) A slug version of a legislative subject,
displayed as \code{url_name} in subject responses.}

\item{key}{your SunlightLabs API key; loads from environment variable from .Renviron or
from an option from .Rprofile}

\item{as}{(character) One of table (default), list, or response (httr response object).
When table is requested, the default, a table is not always returned, but is returned almost
always. You can then make a table yourself if you like.}

\item{...}{Optional additional curl options (debugging tools mostly)
passed on to \code{\link[crul]{HttpClient}}}
}
\value{
some nasty parsing, so just list of lists for now
}
\description{
Get recent bills by subject
}
\examples{
\dontrun{
cg_bills_recent_subject("meat")
}
}
