% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_bills_upcoming.R
\name{cg_bills_upcoming}
\alias{cg_bills_upcoming}
\title{Search for upcoming bills}
\usage{
cg_bills_upcoming(chamber, key = NULL, as = "table", ...)
}
\arguments{
\item{chamber}{The chamber which has scheduled this bill. house or senate}

\item{key}{your SunlightLabs API key; loads from environment variable from .Renviron or
from an option from .Rprofile}

\item{as}{(character) One of table (default), list, or response (httr response object).
When table is requested, the default, a table is not always returned, but is returned almost
always. You can then make a table yourself if you like.}

\item{...}{Optional additional curl options (debugging tools mostly)
passed on to \code{\link[crul]{HttpClient}}}
}
\description{
Search for upcoming bills
}
\examples{
\dontrun{
cg_bills_upcoming("house")
cg_bills_upcoming("senate")
}
}
