% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cg_committees}
\alias{cg_committees}
\title{Gets details (subcommittees + membership) for a committee by id.}
\usage{
cg_committees(member_ids = NULL, committee_id = NULL, chamber = NULL,
  subcommittee = NULL, parent_committee_id = NULL, fields = NULL,
  page = NULL, per_page = NULL, key = getOption("SunlightLabsKey",
  stop("need an API key for Sunlight Labs")), return = "table",
  callopts = list())
}
\arguments{
\item{key}{your SunlightLabs API key; loads from .Rprofile}

\item{return}{(character) One of table (default), list, or response (httr response object)}

\item{callopts}{Optional additional curl options (debugging tools mostly)}

\item{id}{committee id (eg. JSPR)}
}
\value{
Committee details including subcommittees and all members.
}
\description{
Names, IDs, contact info, and memberships of committees and subcommittees in the House and
Senate. All committee information is sourced from bulk data at github.com/unitedstates, which
in turn comes from official House and Senate sources . Feel free to open a ticket with any
bugs or suggestions. We only provide information on current committees and memberships. For
historic data on committee names, IDs, and contact info, refer to the bulk data.
}
\examples{
\dontrun{
cg_committees(member_ids='L000551')
cg_committees(committee_id='SSAP')
cg_committees(committee_id='SSAP', fields='members')
cg_committees(chamber='joint', subcommittee=FALSE)
cg_committees(parent_committee_id='HSWM')

# Disable pagination
cg_committees(per_page='all')

# Output a list
cg_committees(member_ids='L000551', return='list')
# Output an httr response object, for debugging purposes
cg_committees(member_ids='L000551', return='response')
}
}

