% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cg_districts}
\alias{cg_districts}
\title{Get districts for a latitude/longitude or zips}
\usage{
cg_districts(latitude = NULL, longitude = NULL, zip = NULL,
  per_page = 20, page = 1, key = getOption("SunlightLabsKey",
  stop("need an API key for Sunlight Labs")), return = "table", ...)
}
\arguments{
\item{key}{your SunlightLabs API key; loads from .Rprofile}

\item{return}{(character) One of table (default), list, or response (httr response object)}

\item{...}{Optional additional curl options (debugging tools mostly). See examples.}

\item{page}{Page to return. Default: 1. You can use this in combination with the
per_page parameter to get more than the default or max number of results per page.}

\item{per_page}{Number of records to return. Default: 20. Max: 50.}

\item{latitude}{(numeric) latitude of coordinate}

\item{longitude}{(numeric) longitude of coordinate}

\item{zip}{(integer) A 5 digit zip code}
}
\value{
List including data.frame and metadata about results, list, or httr
response object.
}
\description{
Get districts for a latitude/longitude or zips
}
\examples{
\dontrun{
cg_districts(zip = 27511)
cg_districts(latitude = 35.778788, longitude = -78.787805)
}
}

