% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cg_getcommitteesallleg}
\alias{cg_getcommitteesallleg}
\title{Gets a list of all committees that a member serves on, including subcommittes.}
\usage{
cg_getcommitteesallleg(bioguide_id = NULL,
  key = getOption("SunlightLabsKey",
  stop("need an API key for Sunlight Labs")), callopts = list())
}
\arguments{
\item{key}{your SunlightLabs API key; loads from .Rprofile}

\item{callopts}{Optional additional curl options (debugging tools mostly)}

\item{bioguide_id}{legislator's bioguide_id}
}
\value{
Committee details for all committees that the given member serves on.
}
\description{
Gets a list of all committees that a member serves on, including subcommittes.
}
\examples{
\dontrun{
cg_getcommitteesallleg(bioguide_id = 'S000148')
}
}

