% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cg_getdistrictlatlong}
\alias{cg_getdistrictlatlong}
\title{Get districts for a latitude/longitude.}
\usage{
cg_getdistrictlatlong(latitude = NULL, longitude = NULL,
  key = getOption("SunlightLabsKey",
  stop("need an API key for Sunlight Labs")), callopts = list())
}
\arguments{
\item{fields}{You can request specific fields by supplying a vector of fields names. Many fields
are not returned unless requested. If you don’t supply a fields parameter, you will get the
most commonly used subset of fields only. To save on bandwidth, parsing time, and confusion,
it’s recommended to always specify which fields you will be using.}

\item{key}{your SunlightLabs API key; loads from .Rprofile}

\item{return}{(character) One of table (default), list, or response (httr response object)}

\item{...}{Optional additional curl options (debugging tools mostly). See examples.}

\item{page}{Page to return. Default: 1. You can use this in combination with the
per_page parameter to get more than the default or max number of results per page.}

\item{per_page}{Number of records to return. Default: 20. Max: 50.}

\item{latitude}{latitude of coordinate}

\item{longitude}{longitude of coordinate}
}
\value{
List of output fields.
}
\description{
Get districts for a latitude/longitude.
}
\examples{
\dontrun{
cg_getdistrictlatlong(latitude = 35.778788, longitude = -78.787805)
}
}

