% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_nomination.R
\name{cg_nomination}
\alias{cg_nomination}
\title{Get a specific nomination}
\usage{
cg_nomination(congress, id, key = NULL, as = "table", ...)
}
\arguments{
\item{congress}{(integer/numeric) the Congress in which this
nomination was presented}

\item{id}{alphanumeric ID beginning with PN - for example, "PN675"}

\item{key}{your SunlightLabs API key; loads from environment variable from .Renviron or
from an option from .Rprofile}

\item{as}{(character) One of table (default), list, or response (httr response object).
When table is requested, the default, a table is not always returned, but is returned almost
always. You can then make a table yourself if you like.}

\item{...}{Optional additional curl options (debugging tools mostly)
passed on to \code{\link[crul]{HttpClient}}}
}
\description{
Get a specific nomination
}
\examples{
\dontrun{
cg_nomination(115, "PN40")
}
}
\seealso{
Other congress-nominations: \code{\link{cg_nominations_category}},
  \code{\link{cg_nominations_state}}
}
