% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_nominations_state.R
\name{cg_nominations_state}
\alias{cg_nominations_state}
\title{Get nominees by state}
\usage{
cg_nominations_state(congress, state, key = NULL, as = "table", ...)
}
\arguments{
\item{congress}{(integer/numeric) the Congress in which this
nomination was presented}

\item{state}{(character) state Two-letter state abbreviation}

\item{key}{your SunlightLabs API key; loads from environment variable from .Renviron or
from an option from .Rprofile}

\item{as}{(character) One of table (default), list, or response (httr response object).
When table is requested, the default, a table is not always returned, but is returned almost
always. You can then make a table yourself if you like.}

\item{...}{Optional additional curl options (debugging tools mostly)
passed on to \code{\link[crul]{HttpClient}}}
}
\description{
Get nominees by state
}
\examples{
\dontrun{
cg_nominations_state(115, "VA")
}
}
\seealso{
Other congress-nominations: \code{\link{cg_nominations_category}},
  \code{\link{cg_nomination}}
}
