% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw_text.R
\name{cw_text}
\alias{cw_text}
\title{Full-text search the congressional record for instances of a word or phrase}
\usage{
cw_text(...)
}
\description{
Full-text search the congressional record for instances of a word or phrase
}
\keyword{internal}
