% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cw_text}
\alias{cw_text}
\title{Capitol words text.json method. Search the congressional record for instances
of a word or phrase.}
\usage{
cw_text(phrase = NULL, title = NULL, start_date = NULL, end_date = NULL,
  chamber = NULL, state = NULL, party = NULL, bioguide_id = NULL,
  congress = NULL, session = NULL, cr_pages = NULL, volume = NULL,
  page = NULL, key = getOption("SunlightLabsKey",
  stop("need an API key for Sunlight Labs")), callopts = list())
}
\arguments{
\item{phrase}{Phrase to search.}

\item{start_date}{Start date to search on.}

\item{end_date}{End date to search on.}

\item{chamber}{Chamber of congress, House or Senate.}

\item{state}{State, capital two-letter abbreviation (e.g., AK,AZ,NM).}

\item{party}{Political party (one of D,R,I).}

\item{bioguide_id}{Bioguide ID for politician (e.g., B000243)}

\item{key}{Your SunlightLabs API key; loads from .Rprofile.}

\item{callopts}{Further curl options (debugging tools mostly)}

\item{title}{Title of page to search.}

\item{congress}{Congressional session (e.g., 110,111,112)}

\item{session}{Session within the current congress (e.g., 1,2)}

\item{cr_pages}{No definition.}

\item{volume}{No definition.}

\item{page}{No definition.}
}
\value{
Phrases matched.
}
\description{
Capitol words text.json method. Search the congressional record for instances
of a word or phrase.
}
\examples{
\dontrun{
cw_text(phrase='climate change', start_date='2012-09-16', end_date='2012-09-20')
cw_text(phrase='I would have voted', start_date='2011-09-05', end_date='2011-09-16', party='D')
cw_text(phrase='I would have voted', start_date='2011-09-05', end_date='2011-09-16', chamber='House')
cw_text(title='personal explanation', start_date='2011-09-05', end_date='2011-09-16')

out <- cw_text(phrase='climate change', start_date='2010-01-01', end_date='2012-12-01')
out2 <- ldply(2:13, function(x) cw_text(phrase='climate change', start_date='2010-01-01', end_date='2012-12-01', page=x))
alldat <- rbind(out, out2)
str(alldat)
}
}

