% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw_timeseries.R
\name{cw_timeseries}
\alias{cw_timeseries}
\title{Find the popularity of a phrase over a period of time.}
\usage{
cw_timeseries(...)
}
\description{
Find the popularity of a phrase over a period of time.
}
\keyword{internal}
