% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ie_contr_bundled}
\alias{ie_contr_bundled}
\title{Get campaign contribution details}
\usage{
ie_contr_bundled(lobbyist_name = NULL, recipient_name = NULL, page = NULL,
  per_page = NULL, key = getOption("SunlightLabsKey",
  stop("need an API key for Sunlight Labs")), callopts = list())
}
\arguments{
\item{lobbyist_name}{Lobbyist name}

\item{recipient_name}{Recipient name}

\item{page}{The page of results to return; defaults to 1.}

\item{per_page}{The number of results to return per page, defaults to 1,000.
The maximum number of records per page is 100,000.}

\item{key}{your SunlightLabs API key; loads from .Rprofile}

\item{callopts}{Optional additional curl options (debugging tools mostly)}
}
\value{
Details on campaign contributions.
}
\description{
Get campaign contribution details
}
\examples{
\dontrun{
ie_contr_bundled(lobbyist_name='Patton Boggs')
ie_contr_bundled(lobbyist_name='Patton Boggs', per_page=1)
ie_contr_bundled(lobbyist_name='Patton Boggs')
}
}

