% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ie_individuals}
\alias{ie_individuals}
\title{Politician aggregrates: Search for contributions to politicians.}
\usage{
ie_individuals(method = NULL, entity_id = NULL, cycle = NULL,
  limit = NULL, key = getOption("SunlightLabsKey",
  stop("need an API key for Sunlight Labs")), callopts = list())
}
\arguments{
\item{method}{(character) The query string. One of top_ind, top_recorg, top_recpol,
party_breakdown, lobb_reg, lobb_cli, or lobb_iss.}

\item{entity_id}{(character) The transparencydata ID to look up.}

\item{cycle}{(character) Filter results to a particular type of entity. One of politician,
organization, individual or industry.}

\item{limit}{(integer) Limit number of records returned.}

\item{fields}{You can request specific fields by supplying a vector of fields names. Many fields
are not returned unless requested. If you don't supply a fields parameter, you will get the
most commonly used subset of fields only. To save on bandwidth, parsing time, and confusion,
it's recommended to always specify which fields you will be using.}

\item{key}{your SunlightLabs API key; loads from .Rprofile}

\item{return}{(character) One of table (default), list, or response (httr response object)}

\item{...}{Optional additional curl options (debugging tools mostly). See examples.}

\item{page}{Page to return. Default: 1. You can use this in combination with the
per_page parameter to get more than the default or max number of results per page.}

\item{per_page}{Number of records to return. Default: 20. Max: 50.}

\item{order}{Sort results by one or more fields with the order parameter. order is
optional, but if no order is provided, the order of results is not guaranteed to be predictable.
Append \code{__asc} or \code{__desc} to the field names to control sort direction. The default
direction is \code{desc}, because it is expected most queries will sort by a date. Any field
which can be used for filtering may be used for sorting. On full-text search endpoints (URLs
ending in \code{/search}), you may sort by score to order by relevancy.}
}
\value{
A list. Depends on parameters used. Ranges from a single ID returned to basic
   information about the the contributions to and from each entity.
}
\description{
Politician aggregrates: Search for contributions to politicians.
}
\examples{
\dontrun{
# Top individuals
ie_individuals(method='top_ind', limit=1)
ie_individuals(method='top_ind', limit=3)

# Top recipient organizations
ie_individuals(method='top_recorg', entity_id='a03df5d9b20e467fa0ceaefa94c4491e')
ie_individuals(method='top_recorg', entity_id='a03df5d9b20e467fa0ceaefa94c4491e',
   cycle=2012, limit=1)

# Top recipients politicians
ie_individuals(method='top_recpol', entity_id='a03df5d9b20e467fa0ceaefa94c4491e')
ie_individuals(method='top_recpol', entity_id='a03df5d9b20e467fa0ceaefa94c4491e',
   cycle=2012, limit=1)

# Party breakdown
ie_individuals(method='party_breakdown', entity_id='cc768536a9434b9da6fef5846a16ee88')

# Lobbying registrants
ie_individuals(method='lobb_reg', entity_id='4d052c80ef184ce5a5e41e6d34dc452f')

# Lobbying clients
ie_individuals(method='lobb_cli', entity_id='4d052c80ef184ce5a5e41e6d34dc452f', cycle=2002)

# Lobbying issues
ie_individuals(method='lobb_iss', entity_id='4d052c80ef184ce5a5e41e6d34dc452f', limit=2)
}
}

