% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ie_politicians}
\alias{ie_politicians}
\title{Politician aggregrates: Search for contributions to politicians.}
\usage{
ie_politicians(method = NULL, entity_id = NULL, cycle = NULL,
  limit = NULL, key = getOption("SunlightLabsKey",
  stop("need an API key for Sunlight Labs")), callopts = list())
}
\arguments{
\item{method}{(character) The query string. One of top_pol, top_con, unk_ind, top_ind, top_sec,
local_breakdown, type_breakdown, fec_summary, fec_indexp}

\item{entity_id}{(character) The transparencydata ID to look up.}

\item{cycle}{(character) Filter results to a particular type of entity. One of politician,
organization, individual or industry.}

\item{limit}{(integer) Limit number of records returned.}

\item{fields}{You can request specific fields by supplying a vector of fields names. Many fields
are not returned unless requested. If you don't supply a fields parameter, you will get the
most commonly used subset of fields only. To save on bandwidth, parsing time, and confusion,
it's recommended to always specify which fields you will be using.}

\item{key}{your SunlightLabs API key; loads from .Rprofile}

\item{return}{(character) One of table (default), list, or response (httr response object)}

\item{...}{Optional additional curl options (debugging tools mostly). See examples.}

\item{page}{Page to return. Default: 1. You can use this in combination with the
per_page parameter to get more than the default or max number of results per page.}

\item{per_page}{Number of records to return. Default: 20. Max: 50.}

\item{order}{Sort results by one or more fields with the order parameter. order is
optional, but if no order is provided, the order of results is not guaranteed to be predictable.
Append \code{__asc} or \code{__desc} to the field names to control sort direction. The default
direction is \code{desc}, because it is expected most queries will sort by a date. Any field
which can be used for filtering may be used for sorting. On full-text search endpoints (URLs
ending in \code{/search}), you may sort by score to order by relevancy.}
}
\value{
A list. Depends on parameters used. Ranges from a single ID returned to basic
   information about the the contributions to and from each entity.
}
\description{
Politician aggregrates: Search for contributions to politicians.
}
\examples{
\dontrun{
# Top politicians
ie_politicians(method='top_pol', limit=1)
ie_politicians(method='top_pol', limit=3)

# Top contributors
ie_politicians(method='top_con', entity_id='4148b26f6f1c437cb50ea9ca4699417a')
ie_politicians(method='top_con', entity_id='4148b26f6f1c437cb50ea9ca4699417a',
   cycle=2012)

# Top industries
ie_politicians(method='top_ind', entity_id='4148b26f6f1c437cb50ea9ca4699417a')
ie_politicians(method='top_ind', entity_id='4148b26f6f1c437cb50ea9ca4699417a',
   cycle=2012, limit=1)

# Unkown industries
ie_politicians(method='unk_ind', entity_id='4148b26f6f1c437cb50ea9ca4699417a')

# Top sectors
ie_politicians(method='top_sec', entity_id='4148b26f6f1c437cb50ea9ca4699417a')

# Local breakdown
ie_politicians(method='local_breakdown', entity_id='97737bb56b6a4211bcc57a837368b1a4',
   cycle=2002)

# Type breakdown
ie_politicians(method='type_breakdown', entity_id='4148b26f6f1c437cb50ea9ca4699417a')
}
}

