% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{ie_sectors}
\alias{ie_sectors}
\title{Sunlight Labs Influence Explorer API sector code and names to use for searching/retreiving data.}
\description{
Sunlight Labs Influence Explorer API sector code and names to use for searching/retreiving data.
}
\keyword{datasets}

