% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os_committee.R
\name{os_committee}
\alias{os_committee}
\title{Lookup a Committee by ID}
\usage{
os_committee(id, as = "table", key = NULL, ...)
}
\arguments{
\item{id}{committee id (character)}

\item{as}{(character) One of table (default), list, or response}

\item{key}{your SunlightLabs API key; loads from .Rprofile}

\item{...}{Curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\value{
List of output fields.
}
\description{
Lookup a Committee by ID
}
\examples{
\dontrun{
os_committee(id = 'TXC000103')
}
}
