% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os_district.R
\name{os_district}
\alias{os_district}
\title{Get OpenStates districts boundary by boundary id}
\usage{
os_district(id, as = "table", key = NULL, ...)
}
\arguments{
\item{id}{a boundary id used in District Boundary Lookup (character)}

\item{as}{(character) One of table (default), list, or response
(crul response object)}

\item{key}{your SunlightLabs API key; loads from .Rprofile}

\item{...}{Curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\value{
a data.frame of bills.
}
\description{
Get OpenStates districts boundary by boundary id
}
\examples{
\dontrun{
os_district(id = 'ocd-division/country:us/state:tx/sldl:100')
}
}
