% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os_legislator.R
\name{os_legislator}
\alias{os_legislator}
\title{Lookup a Legislator by ID}
\usage{
os_legislator(id, as = "table", key = NULL, ...)
}
\arguments{
\item{id}{legislator id (character)}

\item{as}{(character) One of table (default), list, or response}

\item{key}{your SunlightLabs API key; loads from .Rprofile}

\item{...}{Curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\value{
List of output fields.
}
\description{
Lookup a Legislator by ID
}
\examples{
\dontrun{
os_legislator(id = 'DCL000012')
}
}
