% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os_legislatorgeo.R
\name{os_legislatorgeo}
\alias{os_legislatorgeo}
\title{Lookup Legislators geospatially}
\usage{
os_legislatorgeo(lat, long, as = "table", key = NULL, ...)
}
\arguments{
\item{lat}{a latitude value (numeric,integer)}

\item{long}{a longitude value (numeric,integer)}

\item{as}{(character) One of table (default), list, or response}

\item{key}{your SunlightLabs API key; loads from .Rprofile}

\item{...}{Curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\value{
various, depending on \code{as} parameter
}
\description{
Lookup Legislators geospatially
}
\examples{
\dontrun{
os_legislatorgeo(lat = 35.79, long = -78.78)
}
}
