% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{rsunlight-defunct}
\alias{rsunlight-defunct}
\title{Defunct functions in rsunlight}
\description{
\itemize{
 \item \code{\link{ie_politicians}}: Removed. See \code{Influence Explorer} below
 \item \code{\link{ie_organizations}}: Removed. See \code{Influence Explorer} below
 \item \code{\link{ie_industries}}: Removed. See \code{Influence Explorer} below
 \item \code{\link{ie_individuals}}: Removed. See \code{Influence Explorer} below
 \item \code{\link{ie_lobbying}}: Removed. See \code{Influence Explorer} below
 \item \code{\link{ie_grants}}: Removed. See \code{Influence Explorer} below
 \item \code{\link{ie_contracts}}: Removed. See \code{Influence Explorer} below
 \item \code{\link{ie_faca}}: Removed. See \code{Influence Explorer} below
 \item \code{\link{ie_epa}}: Removed. See \code{Influence Explorer} below
 \item \code{\link{ie_entities}}: Removed. See \code{Influence Explorer} below
 \item \code{\link{ie_earmarks}}: Removed. See \code{Influence Explorer} below
 \item \code{\link{ie_contr_bundled}}: Removed. See \code{Influence Explorer} below
 \item \code{\link{ie_contr}}: Removed. See \code{Influence Explorer} below
}
}
\section{Influence Explorer}{

The Influence Explorer API is now defunct/gone.
}

