% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rsunlight-package.R
\docType{data}
\name{sessions}
\alias{sessions}
\title{Congressional sessions, up to the 114th (2015-2016)}
\description{
Congressional sessions, up to the 114th (2015-2016)
}
\keyword{datasets}

