% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sunlight_df}
\alias{sunlight_df}
\title{Write data from any rsunlight function output to a csv file on your machine.}
\usage{
sunlight_df(x, ...)
}
\arguments{
\item{x}{Output from any of the rsunlight functions.}

\item{...}{Further args passed to read.csv}
}
\description{
Write data from any rsunlight function output to a csv file on your machine.
}
\details{
This function attemps to coerce the raw output from each rsunlight function to a
data.frame to write to csv, but it may fail in some cases. You can always make your own
data.frame.
}
\examples{
\dontrun{
out <- cg_legislators(latitude = 35.778788, longitude = -78.787805)
sunlight_df(sunlight_list(out))
sunlight_df(out)

out <- cg_committees(member_ids='L000551')
sunlight_df(sunlight_list(out))
}
}

