% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{ts_sectors}
\alias{ts_sectors}
\title{Sunlight Labs Transparency API sector code and names to use for searching/retreiving data.}
\description{
Sunlight Labs Transparency API sector code and names to use for searching/retreiving data.
}
\keyword{datasets}

