% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_bill_subjects_search.R
\name{cg_bill_subjects_search}
\alias{cg_bill_subjects_search}
\title{Search for bill subjects that contain a specified term}
\usage{
cg_bill_subjects_search(query, key = NULL, as = "table", ...)
}
\arguments{
\item{query}{(character) a word or phrase to search}

\item{key}{your SunlightLabs API key; loads from environment variable from .Renviron or
from an option from .Rprofile}

\item{as}{(character) One of table (default), list, or response (httr response object).
When table is requested, the default, a table is not always returned, but is returned almost
always. You can then make a table yourself if you like.}

\item{...}{Optional additional curl options (debugging tools mostly)
passed on to \code{\link[crul]{HttpClient}}}
}
\description{
Search for bill subjects that contain a specified term
}
\examples{
\dontrun{
cg_bill_subjects_search('climate')
}
}
