% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_bills.R
\name{cg_bills}
\alias{cg_bills}
\title{Search bills}
\usage{
cg_bills(query, sort = NULL, dir = NULL, key = NULL, as = "table", ...)
}
\arguments{
\item{query}{(character) keyword or phrase. required}

\item{sort}{(character) "_score" or "date" (default)}

\item{dir}{(character) asc or desc (default)}

\item{key}{your SunlightLabs API key; loads from environment variable from .Renviron or
from an option from .Rprofile}

\item{as}{(character) One of table (default), list, or response (httr response object).
When table is requested, the default, a table is not always returned, but is returned almost
always. You can then make a table yourself if you like.}

\item{...}{Optional additional curl options (debugging tools mostly)
passed on to \code{\link[crul]{HttpClient}}}
}
\value{
some nasty parsing, so just list of lists for now
}
\description{
Search bills
}
\examples{
\dontrun{
cg_bills(query = "megahertz")
cg_bills(query = "megahertz", per_page = 3)
}
}
