% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cg_getcommitteeslist}
\alias{cg_getcommitteeslist}
\title{Get list of all committees for a given chamber along with their subcommittees.}
\usage{
cg_getcommitteeslist(chamber = NULL, key = getOption("SunlightLabsKey",
  stop("need an API key for Sunlight Labs")), callopts = list())
}
\arguments{
\item{chamber}{House, Senate, or Joint}

\item{key}{your SunlightLabs API key; loads from .Rprofile}

\item{return}{(character) One of table (default), list, or response (httr response object)}

\item{callopts}{Optional additional curl options (debugging tools mostly)}
}
\value{
List of all committees in the specified chamber with their
   subcommittees (but not memberships due to size of response).
}
\description{
Get list of all committees for a given chamber along with their subcommittees.
}
\examples{
\dontrun{
cg_getcommitteeslist(chamber = 'Joint')
}
}

