% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ie_contr_bundled.R
\name{ie_contr_bundled}
\alias{ie_contr_bundled}
\title{Search for itemized bundlers.}
\usage{
ie_contr_bundled(lobbyist_name = NULL, recipient_name = NULL, page = NULL,
  per_page = NULL, as = "table", key = NULL, ...)
}
\arguments{
\item{lobbyist_name}{Lobbyist name}

\item{recipient_name}{Recipient name}

\item{page}{The page of results to return; defaults to 1.}

\item{per_page}{The number of results to return per page, defaults to 1,000.
The maximum number of records per page is 100,000.}

\item{as}{(character) One of table (default), list, or response (httr response object).
When table is requested, the default, a table is not always returned, but is returned almost
always. You can then make a table yourself if you like.}

\item{key}{your SunlightLabs API key; loads from .Rprofile}

\item{...}{Optional additional curl options (debugging tools mostly)}
}
\value{
A data.frame, list, or httr response object.
}
\description{
Search for itemized bundlers.
}
\examples{
\dontrun{
ie_contr_bundled(lobbyist_name='Patton Boggs')
ie_contr_bundled(lobbyist_name='Patton Boggs', per_page=1)
ie_contr_bundled(lobbyist_name='Patton Boggs')

# most parameters are vectorized, pass in more than one value
ie_contr_bundled(lobbyist_name = c('Patton Boggs', 'Ben Barnes'))
ie_contr_bundled(recipient_name = c('Mitch McConnell', 'Rick Perry'))
}
}

