% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rsunlight-package.R
\docType{data}
\name{ie_sectors}
\alias{ie_sectors}
\title{Sunlight Labs Influence Explorer API sector code and names to use for searching/retreiving data.}
\description{
Sunlight Labs Influence Explorer API sector code and names to use for searching/retreiving data.
}
\keyword{datasets}

