% Generated by roxygen2 (4.0.1): do not edit by hand
\name{os_legislatorsearch}
\alias{os_legislatorsearch}
\title{Search Legislators on OpenStates.}
\usage{
os_legislatorsearch(state = NULL, first_name = NULL, last_name = NULL,
  chamber = NULL, active = NULL, term = NULL, district = NULL,
  party = NULL, key = getOption("SunlightLabsKey",
  stop("need an API key for Sunlight Labs")), callopts = list())
}
\arguments{
\item{key}{your SunlightLabs API key; loads from .Rprofile}

\item{return}{(character) One of table (default), list, or response (httr response object)}

\item{callopts}{Optional additional curl options (debugging tools mostly)}

\item{state}{state two-letter abbreviation (character)}

\item{first_name}{first name of legislator (character)}

\item{last_name}{last name of legislator (character)}

\item{chamber}{one of 'upper' or 'lower' (character)}

\item{active}{TRUE or FALSE (character)}

\item{term}{filter by legislators who served during a certain term (character)}

\item{district}{legislative district (character)}

\item{party}{democratic or republican (character)}
}
\value{
List of output fields.
}
\description{
Search Legislators on OpenStates.
}
\examples{
\dontrun{
os_legislatorsearch(state = 'ca', party = 'democratic')
os_legislatorsearch(state = 'tx', party = 'democratic', active = TRUE)
os_legislatorsearch(state = 'nv', party = 'republican')
os_legislatorsearch(state = 'dc', chamber = 'upper')
}
}

