% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os_statemetasearch.R
\name{os_statemetasearch}
\alias{os_statemetasearch}
\title{Search OpenStates metadata}
\usage{
os_statemetasearch(state = NULL, key = NULL, ...)
}
\arguments{
\item{state}{One or more two-letter state abbreviations (character)}

\item{key}{your SunlightLabs API key; loads from environment variable
from .Renviron or from an option from .Rprofile}

\item{...}{Optional additional curl options (debugging tools mostly)
passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\value{
A list with metadata for each state.
}
\description{
Search OpenStates metadata
}
\examples{
\dontrun{
os_statemetasearch()
os_statemetasearch('ca')
os_statemetasearch('tx')
os_statemetasearch('tx', verbose = TRUE)
}
}
