% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{rsunlight-defunct}
\alias{rsunlight-defunct}
\title{Defunct functions in rsunlight}
\description{
\itemize{
\item \code{ie_politicians}: Removed. See \code{Influence Explorer} below
\item \code{ie_organizations}: Removed. See \code{Influence Explorer} below
\item \code{ie_industries}: Removed. See \code{Influence Explorer} below
\item \code{ie_individuals}: Removed. See \code{Influence Explorer} below
\item \code{ie_lobbying}: Removed. See \code{Influence Explorer} below
\item \code{ie_grants}: Removed. See \code{Influence Explorer} below
\item \code{ie_contracts}: Removed. See \code{Influence Explorer} below
\item \code{ie_faca}: Removed. See \code{Influence Explorer} below
\item \code{ie_epa}: Removed. See \code{Influence Explorer} below
\item \code{ie_entities}: Removed. See \code{Influence Explorer} below
\item \code{ie_earmarks}: Removed. See \code{Influence Explorer} below
\item \code{ie_contr_bundled}: Removed. See \code{Influence Explorer} below
\item \code{ie_contr}: Removed. See \code{Influence Explorer} below
}
}
\section{Influence Explorer}{

The Influence Explorer API is now defunct/gone.
}

