% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{rsunlight-package}
\alias{rsunlight}
\alias{rsunlight-package}
\title{Sunlight Foundation client for R.}
\description{
Sunlight Foundation data from R
}
\details{
You need API keys for Sunlight Foundation APIs. Please get your own API keys if you plant to use
these functions for Sunlight Labs (http://services.sunlightlabs.com/). We set up the functions
so that you can put the key in your .Rprofile file, which will be called on startup of R, and
then you don't have to enter your API key for each run of a function. For example, put this in
your `.Rprofile` file: \code{options(SunlightLabsKey = "YOURKEYHERE")}

Currently we have functions to interface with the following Sunlight Foundation APIs, where
functions for each API are prefixed with a two letter code indicating the service.

\itemize{
 \item Congress API (`cg`)
 \item Open States API (`os`)
 \item Capitol Words API (`cw`)
 \item Influence Explorer API (`ie`)
}

Note that Puerto Rico is not included in Sunlight Foundation data.
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

