% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsunlight-package.R
\docType{package}
\name{rsunlight-package}
\alias{rsunlight-package}
\alias{rsunlight}
\title{Methods for retreiving data from APIs that were
previously under Sulight Labs}
\description{
The various APIs included here are now managed by ProPublica and
OpenStates organizations.
}
\details{
You need API keys for Propublica APIs. Get them at
\url{https://www.propublica.org/datastore/apis}

You need an API key for the OpenStates API. Get it at
\url{https://openstates.org/api/register/}

We set up the functions so that you can use either env vars, or R options.
For env vars, put an entry in your \code{.Renviron} file with the names
\code{PROPUBLICA_API_KEY} and \code{OPEN_STATES_KEY}.

Currently we have functions to interface with the following ProPublica API
and OpenStates API:
\itemize{
\item ProPublica Congress API (\code{cg})
\item Open States API (\code{os})
}

NOTES:
\itemize{
\item Capitol Words API is now defunct
\item We will support ProPublica campaign finance API in the future
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}

Thomas J. Leeper \email{thosjleeper@gmail.com}
}
\keyword{package}
