% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sunlight_tocsv}
\alias{sunlight_tocsv}
\title{Write data to a csv file on your machine.}
\usage{
sunlight_tocsv(x, file = "~/", ...)
}
\arguments{
\item{x}{Output from any of the rsunlight functions.}

\item{file}{File name, with path.}

\item{...}{Further args passed to read.csv}
}
\description{
Write data to a csv file on your machine.
}
\examples{
\dontrun{
out <- cg_getcommittees(id = 'JSPR')
sunlight_tocsv.cg_getcommittees(out, "~/myfile.csv")

out <- cg_getcommitteesallleg(bioguide_id = 'S000148')
sunlight_tocsv.cg_getcommitteesallleg(out, "~/myfile.csv")
}
}

